update of predict file - might contain errors
The tW data set is in time - W0 is a parameter but not estimated so it
was put in auxData.initWeight.tW (set in mydata).

There are three other predictions for weight, but it is weight against
age since birth or age since hatch. So I made two subfunction which
compute weight and structural length according to the age. THe
predictions are different from before so I might have made a mistake.
And I have not yet really debugged this. But the idea is that since we
know the age then the age at stage transitions are known (as are the
sizes) so in the subfunction we compute weight if the age is before
birth, when age is between birth and metamorphosis, etc ....