estimation of f values for control tanks
I changes subfunction for calculatings growth - I put in the dynamic ODE
formulation. We will need this for the tox effects - because we will
want to move a parameter and see how growth is affected ....

I outcommented prections for individual length and weight  because we
don't know what thank they are from.

I put in the pars_init the estimations of the f values for the different
tanks - replicate values seem consistent with each other -

There is a wierd issue I did not solve - I cannnot free all of the 'f'
values at once - the estimation crashes. But I can estimate each 'f'
separately which is waht I did.